Refactor Qt app into dedicated modules
