+++ src/three_dfs/application/__init__.py
@@ -1,0 +1,17 @@
+"""Qt application helpers for the 3dfs desktop shell."""
+
+from .assembly_scanner import (
+    AssemblyRefreshRequest,
+    AssemblyScanOutcome,
+    AssemblyScanWorker,
+    AssemblyScanWorkerSignals,
+)
+from .main_window import MainWindow
+
+__all__ = [
+    "AssemblyRefreshRequest",
+    "AssemblyScanOutcome",
+    "AssemblyScanWorker",
+    "AssemblyScanWorkerSignals",
+    "MainWindow",
+]
